Adding in the comments page to the show page of properties
If you click show on a property now you can see the comment section. You can add a comment which will have a reference to the proprty. Each property will show all of the comments assigned to it when you click show. If you delete a comment it will ask for confirmation. If you delete a property it will also delete all the associated comments from the comment database.